move accessChecker methods to access_checker.go
